Addressing comments received about layout
Signed-off-by: Demetrius Kellum <dmtrs.solutions@gmail.com>

This update allows the Vue File Agent to take up more real estate for
easier dragging and dropping.  Additionally, there's no option to delete
files that have been dropped.  Instead, an error message is displayed
if any of the files are in a non-supported format OR the files are
submitted for rendering if all files are in a supported format.

One issue remains.  When files are dropped on the drop area, it will
expand (briefly) regardless if files are in a supported or non-supported
format.